tests/test_proxy: add test to work with qtest framework with vm connect
Right now the qtest framework provides API to get the VM instance only
by running it. The API added is required to start the qtest application
separately from the VM instance. So VM isn't started by qtest
application and doesn't use the fork() syscall.
For example:
  $ ./tests/test_proxy -q /tmp/qtest-afl.sock -t \
      ./inputfile_000.dump
  $ ./x86_64-softmmu/qemu-system-x86_64 -qtest \
      unix:/tmp/qtest-afl.sock,nowait -qtest-log /tmp/qtest.log -machine \
      accel=qtest -display none -nodefaults -drive \
      if=none,id=drive0,file=/tmp/qtest.raw,format=raw \
      -device virtio-blk-pci,id=drv0,drive=drive0,addr=4.0
In this case the test_proxy application is started and waited for the
qemu connection. After it the inputfile_000.dump file is used as the
input for the virt-blk-queue test case.
This functionality is used as a base for the AFL connection.

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>