virtio-blk: add AFL fuzzing support
The following functionality was added to work with the forkserver AFL
mode:
1. When the request is handled successfully or with an error qemu should
signal AFL that the testcase was handled. It is made by raising the
SIGSTOP signal with the __AFL_LOOP macro defined by afl compiler
wrapper. Put these breakers inside the virtio and virtio-blk code.
2. We should always try to reset the virtual queue state to its initial
state. For now it can be done only by resetting the internal available
ring counters. Maybe in future the descriptor table should be also
reset.
3. In case of packet parsing error qemu is marking device as broken
(hw/virtio/virtio.c:virtio_error()). As a result the device can't handle
the next packet. For fuzzing purposes it is not the case. Don't mark
device as broken.

Separate the code with the AFL_FUZZING macro. This code will be
triggered only if the qemu is build for the virtio-blk fuzzing.

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>