build: add AFL support
Add environment variables to build qemu with AFL. Two environment
variables are used:
 - AFL_PATH: full path to the afl-clang-fast compiler
 - AFL_MODS: the qemu directories which should be built
The build procedure is as follows:
 - Set AFL_PATH
 - Set AFL_MODS, for example AFL_MODS="hw/virtio hw/block"
 - Build qemu
As a result the qemu directories declared in the AFL_MODS variable will
be built with AFL path coverage support.

Add the AFL_FUZZING macro to separate the AFL specific code.

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>